+++ fiddles/react/fiddle-0015-Portfolio/src/App.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+import './App.css';
+import Header from './components/header/header';
+import Footer from './components/footer/footer';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <Header/>
+        <Footer/>
+      </div>
+    );
+  }
+}
+
+export default App;